requirements/base.txt
Django==1.9
django-suit==0.2.15
django-braces==1.8.1
mysqlclient==1.3.7
python-social-auth==0.2.13
github3.py==0.9.4
